Full rss feed (#156)
* Add feed configuration to _config.yml

This commit introduces a new feed configuration with a posts limit of 10 and sets excerpt_only to false, enhancing the blog's feed functionality.

* Add jekyll-feed plugin to Gemfile and update configuration

This commit adds the jekyll-feed gem to the Gemfile and includes a new plugins section in _config.yml to enable the feed functionality, enhancing the blog's capabilities for post syndication.